backend/pyproject.toml
[project]
name = "itsfriday"
version = "1.0.0"
description = "ItsFriday - API Observability Platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "ItsFriday Team" }
]

dependencies = [
    "django>=5.0,<6.0",
    "django-ninja>=1.0,<2.0",
    "django-cors-headers>=4.0,<5.0",
    "psycopg[binary]>=3.1,<4.0",
    "clickhouse-driver>=0.2.6,<1.0",
    "celery>=5.3,<6.0",
    "redis>=5.0,<6.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "gunicorn>=21.0,<23.0",
    "httpx>=0.25,<1.0",
    "sentry-sdk[django]>=2.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-django>=4.5,<5.0",
    "pytest-asyncio>=0.21,<1.0",
    "black>=23.0,<25.0",
    "ruff>=0.1,<1.0",
    "mypy>=1.5,<2.0",
    "django-stubs>=4.2,<6.0",
    "ipython>=8.0,<9.0",
]

prod = [
    "sentry-sdk[django]>=1.29,<3.0",
    "boto3>=1.28,<2.0",
    "django-storages>=1.14,<2.0",
]

test = [
    "pytest>=7.4,<9.0",
    "pytest-django>=4.5,<5.0",
    "pytest-cov>=4.1,<6.0",
    "pytest-asyncio>=0.21,<1.0",
    "factory-boy>=3.3,<4.0",
    "faker>=19.0,<31.0",
    "httpx>=0.25,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["src", "config", "api", "apps", "core"]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "migrations",
    ".venv",
]

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["src/tests"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/config/settings/*",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
